* updates fits_file.h to write true datacubes * completely replaces old-style fits writing with Reed's fits engine * fixes incorrect datatypes * removes a lot of debug logging

+++ utils/daemonize.h
@@ -70,7 +70,7 @@ namespace Daemon {
 
     // Close all open file descriptors
     //
-    for( int fd = sysconf( _SC_OPEN_MAX ); fd >= 0; fd-- ) close( fd );
+    for( long fd = sysconf( _SC_OPEN_MAX ); fd >= 0; fd-- ) close( static_cast<int>(fd) );
 
     // reopen stdin, stdout, stderr
     //
